pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cricket-intelligence"
version = "0.1.0"
description = "AI-powered cricket intelligence system with stats and news"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "Cricket Intelligence Team"}
]

dependencies = [
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    "langgraph>=0.0.20",
    "openai>=1.12.0",
    "duckdb>=0.9.2",
    "chromadb>=0.4.22",
    "sentence-transformers>=2.3.1",
    "pandas>=2.1.4",
    "pydantic>=2.5.3",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "mcp>=0.1.0",
    "requests>=2.31.0",
    "sqlparse>=0.4.4",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "pytest-asyncio>=0.21.1",
    "black>=23.12.1",
    "ruff>=0.1.9",
    "mypy>=1.8.0",
    "ipykernel>=6.29.0",
    "jupyter>=1.0.0",
]

[project.scripts]
cricket-agent = "cricket_intelligence.agent.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
]
markers = [
    "unit: Unit tests (fast, no external dependencies)",
    "integration: Integration tests (require database)",
    "e2e: End-to-end tests (full workflow)",
    "slow: Slow tests (skip with -m 'not slow')",
]

[tool.black]
line-length = 100
target-version = ["py310"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | data
)/
'''

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    "data",
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false  # Set to true for stricter typing
ignore_missing_imports = true
